requirements.txt
vk_api
pydub
psycopg2-binary
environs
yt_dlp
